# Faster 256-bit multiplication