--- a/tests/performance/bigint_arithm.xml
+++ b/tests/performance/bigint_arithm.xml
@@ -8,8 +8,8 @@
     <query>SELECT toInt128(number) * number FROM numbers_mt(1000000000) FORMAT Null</query>
     <query>SELECT toInt128(number) / number FROM numbers_mt(1000000000) FORMAT Null</query>
 
-    <query>SELECT toInt256(number) + number FROM numbers_mt(1000000000) FORMAT Null</query>
-    <query>SELECT toInt256(number) - number FROM numbers_mt(1000000000) FORMAT Null</query>
-    <query>SELECT toInt256(number) * number FROM numbers_mt(1000000000) FORMAT Null</query>
-    <query>SELECT intDiv(toInt256(number + 1), number + 1) FROM numbers_mt(1000000000) FORMAT Null</query>
+    <query>SELECT toInt256(number) + number FROM numbers_mt(100000000) FORMAT Null</query>
+    <query>SELECT toInt256(number) - number FROM numbers_mt(100000000) FORMAT Null</query>
+    <query>SELECT toInt256(number) * number FROM numbers_mt(100000000) FORMAT Null</query>
+    <query>SELECT intDiv(toInt256(number + 1), number + 1) FROM numbers_mt(100000000) FORMAT Null</query>
 </test>