--- a/tests/queries/0_stateless/01485_256_bit_multiply.sql
+++ b/tests/queries/0_stateless/01485_256_bit_multiply.sql
@@ -1,0 +1,4 @@
+select count() from
+(
+    select toInt128(number) * number x, toInt256(number) * number y from numbers_mt(100000000) where x != y
+);