… tests/queries/0_stateless/01485_256_bit_multiply.reference