I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement

Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Faster 256-bit multiplication.

Detailed description / Documentation draft:
The idea is form https://github.com/abseil/abseil-cpp/blob/master/absl/numeric/int128.h `operator*`

```
SELECT toInt128(number) * number
FROM numbers_mt(1000000000)
FORMAT Null

0 rows in set. Elapsed: 0.372 sec. Processed 1.00 billion rows, 8.00 GB (2.69 billion rows/s., 21.48 GB/s.) 
```
```
SELECT toInt256(number) * number
FROM numbers_mt(1000000000)
FORMAT Null

0 rows in set. Elapsed: 2.999 sec. Processed 1.00 billion rows, 8.00 GB (333.44 million rows/s., 2.67 GB/s.) 
```